refactor: optimize images, convert to `webp`
