+++ index.d.ts
@@ -1,5 +1,2 @@
-declare module "*.png"
-declare module "*.jpg"
-declare module "*.svg"
 declare module "*.webp"
 declare module "react-animated-cursor"
