+++ src/components/Common/Header.tsx
@@ -2,7 +2,7 @@ import type { FC } from "react"
 import { useEffect, useRef,useState } from "react"
 import { Link } from "react-router-dom"
 
-import Logo from "../../assets/images/Logo.png"
+import Logo from "../../assets/images/Logo.webp"
 
 import MenuButton from "./MenuButton"
 import NavLink from "./NavLink"
